requirements.txt
python-dotenv~=1.0.1
langchain-community~=0.3.24
langchain~=0.3.25
requests~=2.32.5
notion-client~=2.5.0
dashscope~=1.22.2
streamlit~=1.49.1
openai~=1.107.2
langchain-core~=0.3.75
llama-index~=0.14.3
faiss-cpu~=1.12.0